c++: convert_to_void and volatile references
Jason pointed out that even implicit loads of volatile references need
to undergo lvalue-to-rvalue conversion, but we currently emit a warning
in this case and discard the load.  This patch changes this behavior so
that we don't issue a warning, and preserve the load.

gcc/cp/ChangeLog:

	* cvt.cc (convert_to_void) <case INDIRECT_REF>: Remove warning
	for an implicit load of a volatile reference.  Simplify as if
	is_reference is false.  Check REFERENCE_REF_P in the test
	guarding the -Wunused-value diagnostic.

gcc/testsuite/ChangeLog:

	* g++.dg/expr/discarded1a.C: No longer expect warning for
	implicit load of a volatile reference.
	* g++.old-deja/g++.bugs/900428_01.C: Likewise.
	* g++.dg/expr/volatile2.C: New test.