Darwin, debug : Switch to DWARF 3 or 4 when dsymutil supports it.
Tested on i686, x86_64 and aarch64 Darwin, also on x86_64 and aarch64
Linux.  Having said this, some fallout might well be expected on Darwin
since the consumers have not had much use with GCC output where the
DWARF version is > 2, we will have to tackle that as it arises.
pushed to trunk, thanks,

Iain