target/30484 - testcase for exploration
The following adds a testcase for the unfixed PR, skipped on the
known affected target.  The intent is to discover targets that
suffer from the same issue or whose sdiv do not wrap as expected
with -fwrapv.

Tested on x86_64-unknown-linux-gnu.

OK?  Jeff, do you want to put this on your tester first (do they
perform runtime simulator testing?)

	PR target/30484
	* gcc.dg/torture/pr30484.c: New testcase.