librust: Add libproc_macro and build system
This patch series adds the build system changes to allow the Rust
frontend to develop and distribute its own libraries. The first library
we have been working on is the `proc_macro` library, comprised of a C++
library as well as a user-facing Rust library.

Follow up commits containing the actual library code will be committed.
Should I submit patches to the MAINTAINERS file to allow Philip and I to
commit to this folder without first approval?

This first commit adds a simple `libgrust` folder with on top of which the
full library will be built.

All the best,

Arthur

-----

Add some dummy files in libproc_macro along with it's build system.

ChangeLog:

	* libgrust/Makefile.am: New file.
	* libgrust/configure.ac: New file.
	* libgrust/libproc_macro/Makefile.am: New file.
	* libgrust/libproc_macro/proc_macro.cc: New file.
	* libgrust/libproc_macro/proc_macro.h: New file.

Signed-off-by: Pierre-Emmanuel Patry <pierre-emmanuel.patry@embecosm.com>