c++: constraint rewriting during ttp coercion [PR111485]
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk and perhaps backports?

-- >8 --

In order to compare the constraints of a ttp with that of its argument,
we rewrite the ttp's constraints in terms of the argument template's
template parameters.  The substitution to achieve this currently uses a
single level of template arguments, but that never does the right thing
because a ttp's template parameters always have level >= 2.  This patch
fixes this by including the outer template arguments in this substitution,
which ought to match the depth of the ttp.

The second testcase demonstrates that it's sometimes necessary to
substitute the concrete outer template arguments instead of generic
ones, because a ttp's constraints could depend on outer arguments.

	PR c++/111485

gcc/cp/ChangeLog:

	* pt.cc (is_compatible_template_arg): New parameter 'args'.
	Add the outer template arguments 'args' to 'new_args'.
	(convert_template_argument): Pass 'args' to
	is_compatible_template_arg.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/concepts-ttp5.C: New test.
	* g++.dg/cpp2a/concepts-ttp6.C: New test.