light expander sra
Hi,

There are a few PRs (meta-bug PR101926) on various targets.
The root causes of them are similar: the aggeragte param/
returns are passed by multi-registers, but they are
stored to stack from registers first; and then, access the
parameter through stack slot.

This would lead to generating suboptimal instructions,
because rtl passes may be not able to optimize it.

While a general idea, it would be better to access the
aggregate parameters/returns directly through registers
and get faster code.  This idea would be a kind of SRA
(using the scalar registers to access the aggregate
parameter/returns).

Adopting an sra for parameters/returns in the expander
would be practical.
For aggregate parameters that passed via multi-registers,
the INCOMING_RTL contains a list of the "incoming hard
registers" in a parallel rtx (or using the first register
and total size to indicate the list of registers).
It is similar for returns, the outgoing registers are
recorded in DECL_RTL of DECL_RESULT for the function.
When accessing the aggregate parm/return, we could figure
out which parts of incoming/outgoing are accessed.

The initial implementation for this sra (light-expander-sra)
contains below parts:

a. Check if the aggregate parameters/returns are ok/profitable
  to scalarize the accesses, and set the scalar pseudos for
  the aggregate parameter/return.
  - This is done in "expand_function_start", after the
    incoming/outgoing hard registers are determined for the
    paramter(s)/return.
    Now, the scalarized registers are recorded in DECL_RTL
    for the parameter/return in parallel form.
  - When set the DECL_RTL, "scalarizable_aggregate" is called
    to check the accesses are ok/profitable to scalarize.
    To support more cases, we can enhance this function.
    For example:
    - 'reverse storage order' can be supported.
    - 'TImode/vector-mode from multi-regs' may be supported.
    - some cases on 'writing to parameter'/'overlap accesses'
      maybe ok to support.

b. When expanding the accesses on the parameters/returns,
  according to the access info (e.g. bitpos,bitsize, mode),
  compute the scalars(pseudos), and use the pseudos to
  expand the access.  For example:
  - Expand the component access of a parameter: "_1 = arg.f1"
    or the access on a whole parameter: rhs of "_2 = arg"
  - Expand the assignment to a return val:
    "D.xx = yy; or D.xx.f = zz" where D.xx occurs on return
    stmt.
  - This is mainly done in expr.cc(expand_expr_real_1,
    expand_assignment).  To expand accesses, function
    "extract_sub_member" is used to compute the scalar rtx.

Except above two parts, some work is done in the GIMPLE tree:
collect sra candidates for parameters/returns, and collect
the SRA access info.
This is mainly done at the beginning of the expander pass by
the class (named expand_sra) and its member functions.
Below are two items of this part.
 - Collect light-expand-sra candidates.
  Each parameter is checked if it has the proper aggregate type.
  Collect return val (VAR_P) on each return stmts if the function
  is returning via registers.
  This is implemented in expand_sra::collect_sra_candidates.

 - Build/collect/manage all the access on the candidates.
  The function "scan_function" is used to do this work, it goes
  through all basicblocks and all interesting stmts (phi, return,
  assign, call, asm) are checked.
  If there is an interesting expression (e.g. COMPONENT_REF or
  PARM_DECL) on the aggregate parameter, record the required info
  for the access (e.g. pos, size, type, base).
  For some access, it may be not possible to use scalar registers.
  e.g.: the aggregate parameter is address-taken and
  accessed via memory. "foo(struct S arg) {bar (&arg);}"

Another thing about this patch: trying to common code for
light-expand-sra, tree-sra, and ipa-sra.
 - Now, the class "expand_sra" is based on "sra_default_analyzer".
  "sra_default_analyzer" provides stmt analyzing interfaces.
 - "scan_function" provides the capability to go through functions
  and check interesting sra stmts.
 - We can continue refactoring the code to share similar frameworks
  even if the behaviors are slightly different between those SRAs.

This patch is tested on ppc64{,le}.
Thanks a lot for your review in advance?

BR,
Jeff (Jiufu Guo)

	PR target/65421

gcc/ChangeLog:

	* cfgexpand.cc (struct access): New class.
	(struct expand_sra): New class.
	(expand_sra::collect_sra_candidates): New member function.
	(expand_sra::add_sra_candidate): Likewise.
	(expand_sra::build_access): Likewise.
	(expand_sra::analyze_phi): Likewise.
	(expand_sra::analyze_assign): Likewise.
	(expand_sra::visit_base): Likewise.
	(expand_sra::protect_mem_access_in_stmt): Likewise.
	(expand_sra::expand_sra):  Class constructor.
	(expand_sra::~expand_sra): Class destructor.
	(expand_sra::scalarizable_access):  New member function.
	(expand_sra::scalarizable_accesses):  Likewise.
	(scalarizable_aggregate):  New function.
	(set_scalar_rtx_for_returns):  New function.
	(expand_value_return): Updated.
	(expand_debug_expr): Updated.
	(pass_expand::execute): Updated to use expand_sra.
	* cfgexpand.h (scalarizable_aggregate): New declare.
	(set_scalar_rtx_for_returns): New declare.
	* expr.cc (expand_assignment): Updated.
	(expand_constructor): Updated.
	(query_position_in_parallel): New function.
	(extract_sub_member): New function.
	(expand_expr_real_1): Updated.
	* expr.h (query_position_in_parallel): New declare.
	* function.cc (assign_parm_setup_block): Updated.
	(assign_parms): Updated.
	(expand_function_start): Updated.
	* tree-sra.h (struct sra_base_access): New class.
	(struct sra_default_analyzer): New class.
	(scan_function): New template function.
	* var-tracking.cc (track_loc_p): Updated.

gcc/testsuite/ChangeLog:

	* g++.target/powerpc/pr102024.C: Updated
	* gcc.target/powerpc/pr108073.c: New test.
	* gcc.target/powerpc/pr65421-1.c: New test.
	* gcc.target/powerpc/pr65421-2.c: New test.