Introduce -finline-stringops
On Sep 21, 2023, Alexandre Oliva <oliva@adacore.com> wrote:

> On Sep 15, 2023, Alexandre Oliva <oliva@adacore.com> wrote:
>> On Jun 22, 2023, Alexandre Oliva <oliva@adacore.com> wrote:
>>> On Jun  2, 2023, Alexandre Oliva <oliva@adacore.com> wrote:
>>>> Introduce -finline-stringops

>>> Ping?  https://gcc.gnu.org/pipermail/gcc-patches/2023-June/620472.html

>> Ping?

> Here's a refreshed and improved patch, that improves the memcmp
> expansion a little, dropping redundant barriers and conditional
> branches.

I was advised by linaro's CI that there are some failures to inline
memcpy (bcopy, really) in the memmove test, that only required memmove
inlining.  Oops.  That was a mistake in the testcase.  There was no
reason to demand memcpy to be inlined.

While looking into it, I found another case in which the memset expander
could loop forever.  The assumption that can_store_by_pieces would
return true at least for 1-byte stores doesn't hold on aarch64.  A
1-byte store only expands with setmem, not with store_by_pieces.

I had expected store_by_pieces to always pass for QImode constants, and
perhaps also for other natural machine modes, but...  I guess I can't
count on that.  I see aarch64's SET_RATIO and CLEAR_RATIO are sometimes
set to zero, so as to disable the by-pieces machinery even for
single-instruction stores, and that setmem actually outputs better code
than by_pieces for all but single-byte stores, and equivalent code for
single-byte stores.

So it looks like integrating setmem into by-multiple-pieces would be the
way to go, but...  there's no way to test whether setmem is available
for a certain combination of operands, as there is for store_by_pieces.
So, given that -finline-stringops is not so much about performance as it
is about avoiding runtime dependencies, I'm only making sure we have a
fallback to use for the single-byte case.

I suppose I could introduce some target hook for machines to test for
supported setmem operands without compile-time rtl allocation, and use
that if defined, but it could be cumbersome to introduce only to improve
performance in scenarios where performance is not critical.

Regstrapped on x86_64-linux-gnu.  Also tested on an aarch64-elf target.
Ok to install?

Introduce -finline-stringops

try_store_by_multiple_pieces was added not long ago, enabling
variable-sized memset to be expanded inline when the worst-case
in-range constant length would, using conditional blocks with powers
of two to cover all possibilities of length and alignment.

This patch introduces -finline-stringops[=fn] to request expansions to
start with a loop, so as to still take advantage of known alignment
even with long lengths, but without necessarily adding store blocks
for every power of two.

This makes it possible for the supported stringops (memset, memcpy,
memmove, memset) to be expanded, even if storing a single byte per
iteration.  Surely efficient implementations can run faster, with a
pre-loop to increase alignment, but that would likely be excessive for
inline expansions.

Still, in some cases, such as in freestanding environments, users
prefer to inline such stringops, especially those that the compiler
may introduce itself, even if the expansion is not as performant as a
highly optimized C library implementation could be, to avoid
depending on a C runtime library.

for  gcc/ChangeLog

	* expr.cc (emit_block_move_hints): Take ctz of len.  Obey
	-finline-stringops.  Use oriented or sized loop.
	(emit_block_move): Take ctz of len, and pass it on.
	(emit_block_move_via_sized_loop): New.
	(emit_block_move_via_oriented_loop): New.
	(emit_block_move_via_loop): Take incr.  Move an incr-sized
	block per iteration.
	(emit_block_cmp_via_cmpmem): Take ctz of len.  Obey
	-finline-stringops.
	(emit_block_cmp_via_loop): New.
	* expr.h (emit_block_move): Add ctz of len defaulting to zero.
	(emit_block_move_hints): Likewise.
	(emit_block_cmp_hints): Likewise.
	* builtins.cc (expand_builtin_memory_copy_args): Pass ctz of
	len to emit_block_move_hints.
	(try_store_by_multiple_pieces): Support starting with a loop.
	(expand_builtin_memcmp): Pass ctz of len to
	emit_block_cmp_hints.
	(expand_builtin): Allow inline expansion of memset, memcpy,
	memmove and memcmp if requested.
	* common.opt (finline-stringops): New.
	(ilsop_fn): New enum.
	* flag-types.h (enum ilsop_fn): New.
	* doc/invoke.texi (-finline-stringops): Add.

for  gcc/testsuite/ChangeLog

	* gcc.dg/torture/inline-mem-cmp-1.c: New.
	* gcc.dg/torture/inline-mem-cpy-1.c: New.
	* gcc.dg/torture/inline-mem-cpy-cmp-1.c: New.
	* gcc.dg/torture/inline-mem-move-1.c: New.
	* gcc.dg/torture/inline-mem-set-1.c: New.